+++ public/assets/architecture-7GRP2DOG-CRV6iB3M.js
@@ -1,0 +1,1 @@
+import"./chunk-FOHPRMQF-DHwB1DNv.js";import{M as e}from"./mermaid-parser.core-BxAKQ5vj.js";export{e as createArchitectureServices};